+++ src/database.py
@@ -1,7 +1,7 @@
 from datetime import date
 import sqlite3
 from sqlite3 import Error
-from users import roles
+from roles import roles
 from cryptoUtils import cryptoUtils
 import os
