+++ src/roles.py
@@ -1,0 +1,5 @@
+from enum import Enum
+class roles(Enum):
+    CONSULTANT = 'consultant'
+    ADMIN = 'admin'
+    SUPERADMIN = 'superadmin'
