+++ src/userBlueprint.py
@@ -1,0 +1,5 @@
+class userBlueprint:
+    def __init__(self, id, userName,db):
+        self.id = id
+        self.userName = userName
+        self.db = db
