+++ src/userInterface.py
@@ -2,7 +2,7 @@ from users import roles
 from users import consultant
 from users import systemAdministrator
 from users import superAdministrator
-from users import roles
+from roles import roles
 import os
 import time
